+++ crates/ruma-identifiers-validation/src/event_id.rs
@@ -2,7 +2,7 @@ use crate::{validate_delimited_id, Error};
 
 pub fn validate(s: &str) -> Result<(), Error> {
     if s.contains(':') {
-        validate_delimited_id(s, &['$'])?;
+        validate_delimited_id(s, b'$')?;
     } else if !s.starts_with('$') {
         return Err(Error::MissingLeadingSigil);
     }