+++ crates/ruma-identifiers-validation/src/room_alias_id.rs
@@ -1,5 +1,5 @@
 use crate::{validate_delimited_id, Error};
 
 pub fn validate(s: &str) -> Result<(), Error> {
-    validate_delimited_id(s, &['#'])
+    validate_delimited_id(s, b'#')
 }